pyproject.toml
[project]
name = "glee-back"
version = "0.1.0"
description = ""
authors = [
    {name = "sungchan1",email = "sungchan254@naver.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "black (>=25.1.0,<26.0.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "ruff (>=0.9.7,<0.10.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "pydantic-settings (>=2.8.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "motor (>=3.7.0,<4.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pillow (>=11.1.0,<12.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.mypy]
plugins = "pydantic.mypy"
python_version = "3.11"
strict = true
exclude = 'airflow/.*'

[tool.pytest.ini_options]
asyncio_mode = "auto"
